limits: Make max_vertex_attribs an unsigned

+++ src/gl_limits.rs
@@ -1,16 +1,15 @@
-use gleam::gl::types::GLint;
 use gleam::gl;
 
 #[derive(Debug, Clone)]
 #[cfg_attr(feature="serde_serialization", derive(Serialize, Deserialize))]
 pub struct GLLimits {
-    pub max_vertex_attribs: GLint,
+    pub max_vertex_attribs: u32,
 }
 
 impl GLLimits {
     pub fn detect() -> GLLimits {
         GLLimits {
-            max_vertex_attribs: gl::get_integer_v(gl::MAX_VERTEX_ATTRIBS),
+            max_vertex_attribs: gl::get_integer_v(gl::MAX_VERTEX_ATTRIBS) as u32,
         }
     }
 }
